+++ src/components/uikit/index.tsx
@@ -1,7 +1,6 @@
 export { default as AddButton } from './AddButton';
 export { default as DatePicker } from './DatePicker';
 export { default as DeleteButton } from './DeleteButton';
-export { default as EditTransactionModal } from './EditTransactionModal';
 export { default as GenericButton } from './GenericButton';
 export { default as GenericModal } from './GenericModal';
 export { default as InputInteger } from './InputInteger';
@@ -13,7 +12,6 @@ export { default as ModalInform } from './ModalInform';
 export { default as SaveButton } from './SaveButton';
 export { default as SelectMenu } from './SelectMenu';
 export { default as TextInput } from './TextInput';
-export { default as AddTransactionModal } from './AddTransactionModal';
 export { default as SelectPayer } from './SelectPayer';
 export { default as GroupCategoryInput } from './GroupCategoryInput';
 export { default as SelectCompleteFlag } from './SelectCompleteFlag';
